+++ R/sepList.R
@@ -63,7 +63,6 @@
 ##'  cat("Objects in the local environment after separating the list:\n")
 ##'  print(ls(.GlobalEnv))
 ##'
-##'
 ##'} # sepTest
 ##'
 ##'sepTest(list(z1 = 10, z2 = "that"))